+++ src/pages/Content/modules/components/content-loader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ContentLoader';
