+++ src/pages/Content/modules/components/skeleton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Skeleton';
