+++ src/pages/Content/modules/components/skeleton/stories/skeleton.stories.tsx
@@ -1,0 +1,14 @@
+/* eslint-disable react/function-component-definition */
+import { ComponentMeta, ComponentStory } from '@storybook/react';
+import React from 'react';
+import Skeleton from '../Skeleton';
+
+export default { title: 'Components/Skeleton' } as ComponentMeta<
+  typeof Skeleton
+>;
+
+export const SkeletonLoader: ComponentStory<typeof Skeleton> = () => (
+  <div style={{ width: '288px', padding: '24px' }}>
+    <Skeleton />
+  </div>
+);
